requirements.txt
dbt-postgres
dbt-core
psycopg2-binary
python-dotenv